+++ ams/models/group.py
@@ -1,7 +1,12 @@
-import logging  # NOQA
+import logging
+
+import numpy as np
 
 from andes.models.group import GroupBase as andes_GroupBase
 from andes.core.service import BackRef
+from andes.utils.func import validate_keys_values
+
+from ams.shared import pd
 
 logger = logging.getLogger(__name__)
 
@@ -83,6 +88,165 @@ class GroupBase(andes_GroupBase):
 
         model = self.idx2model(to_idx)
         model.set_backref(name, from_idx, to_idx)
+
+    def alter(self, src, idx, value, attr='v'):
+        """
+        Alter values of input parameters or constant service for a group of models.
+
+        .. note::
+            New in version 0.9.14. Duplicate of `andes.models.group.GroupBase.alter`.
+
+        Parameters
+        ----------
+        src : str
+            The parameter name to alter
+        idx : str, float, int
+            The unique identifier for the device to alter
+        value : float
+            The desired value
+        attr : str, optional
+            The attribute to alter. Default is 'v'.
+        """
+        self._check_src(src)
+        self._check_idx(idx)
+
+        idx, _ = self._1d_vectorize(idx)
+        models = self.idx2model(idx)
+
+        if isinstance(value, (str, int, float, np.integer, np.floating)):
+            value = [value] * len(idx)
+
+        for mdl, ii, val in zip(models, idx, value):
+            mdl.alter(src, ii, val, attr=attr)
+
+        return True
+
+    def as_dict(self, vin=False):
+        """
+        Export group common parameters as a dictionary.
+
+        .. note::
+            New in version 0.9.14. Duplicate of `andes.models.group.GroupBase.as_dict`.
+
+        This method returns a dictionary where the keys are the `Model` parameter names
+        and the values are array-like structures containing the data in the order they were added.
+        Unlike `Model.as_dict()`, this dictionary does not include the `uid` field.
+
+        Parameters
+        ----------
+        vin : bool, optional
+            If True, includes the `vin` attribute in the dictionary. Default is False.
+
+        Returns
+        -------
+        dict
+            A dictionary of common parameters.
+        """
+        out_all = []
+        out_params = self.common_params.copy()
+        out_params.insert(2, 'idx')
+
+        for mdl in self.models.values():
+            if mdl.n <= 0:
+                continue
+            mdl_data = mdl.as_df(vin=True) if vin else mdl.as_dict()
+            mdl_dict = {k: mdl_data.get(k) for k in out_params if k in mdl_data}
+            out_all.append(mdl_dict)
+
+        if not out_all:
+            return {}
+
+        out = {key: np.concatenate([item[key] for item in out_all]) for key in out_all[0].keys()}
+        return out
+
+    def as_df(self, vin=False):
+        """
+        Export group common parameters as a `pandas.DataFrame` object.
+
+        .. note::
+            New in version 0.9.14. Duplicate of `andes.models.group.GroupBase.as_df`.
+
+        Parameters
+        ----------
+        vin : bool
+            If True, export all parameters from original input (``vin``).
+
+        Returns
+        -------
+        DataFrame
+            A dataframe containing all model data. An `uid` column is added.
+        """
+        return pd.DataFrame(self.as_dict(vin=vin))
+
+    def find_idx(self, keys, values, allow_none=False, default=None, allow_all=False):
+        """
+        Find indices of devices that satisfy the given `key=value` condition.
+
+        This method iterates over all models in this group.
+
+        .. note::
+            New in version 0.9.14. Duplicate of `andes.models.group.GroupBase.find_idx`.
+
+        Parameters
+        ----------
+        keys : str, array-like, Sized
+            A string or an array-like of strings containing the names of parameters for the search criteria.
+        values : array, array of arrays, Sized
+            Values for the corresponding key to search for. If keys is a str, values should be an array of
+            elements. If keys is a list, values should be an array of arrays, each corresponding to the key.
+        allow_none : bool, optional
+            Allow key, value to be not found. Used by groups. Default is False.
+        default : bool, optional
+            Default idx to return if not found (missing). Default is None.
+        allow_all : bool, optional
+            Return all matches if set to True. Default is False.
+
+        Returns
+        -------
+        list
+            Indices of devices.
+        """
+
+        keys, values = validate_keys_values(keys, values)
+
+        n_mdl, n_pair = len(self.models), len(values[0])
+
+        indices_found = []
+        # `indices_found` contains found indices returned from all models of this group
+        for model in self.models.values():
+            indices_found.append(model.find_idx(keys, values, allow_none=True, default=default, allow_all=True))
+
+        # --- find missing pairs ---
+        i_val_miss = []
+        for i in range(n_pair):
+            idx_cross_mdls = [indices_found[j][i] for j in range(n_mdl)]
+            if all(item == [default] for item in idx_cross_mdls):
+                i_val_miss.append(i)
+
+        if (not allow_none) and i_val_miss:
+            miss_pairs = []
+            for i in i_val_miss:
+                miss_pairs.append([values[j][i] for j in range(len(keys))])
+            raise IndexError(f'{keys} = {miss_pairs} not found in {self.class_name}')
+
+        # --- output ---
+        out_pre = []
+        for i in range(n_pair):
+            idx_cross_mdls = [indices_found[j][i] for j in range(n_mdl)]
+            if all(item == [default] for item in idx_cross_mdls):
+                out_pre.append([default])
+                continue
+            for item in idx_cross_mdls:
+                if item != [default]:
+                    out_pre.append(item)
+                    break
+
+        if allow_all:
+            out = out_pre
+        else:
+            out = [item[0] for item in out_pre]
+
+        return out
 
 
 class Undefined(GroupBase):